Updated Container to support .extend (just export the styled component)
