+++ src/components/Header/styled.js
@@ -5,6 +5,6 @@ export const Wrapper = styled.header`
   background-color: white;
   display: flex;
   justify-content: space-between;
-  padding: 30px ${gridColWidth / 2}px;
+  margin: 30px ${gridColWidth / 2}px;
   align-items: center;
 `;
