+++ src/compositions/pages/Home/BuildBetter/index.jsx
@@ -1,11 +1,10 @@
 import React from 'react';
 
 import Button from '../../../../components/Button';
-import Container from '../../../../components/Container';
-import { Body, Title } from './styled';
+import { Body, Section, Title } from './styled';
 
 const BuildBetter = () => (
-  <Container>
+  <Section>
     <Title>Build better</Title>
     <Body>
       Great companies go beyond their customers expectations, over and over
@@ -17,7 +16,7 @@ const BuildBetter = () => (
       experiences.
     </Body>
     <Button href="#case-studies">Case Studies ↓</Button>
-  </Container>
+  </Section>
 );
 
 export default BuildBetter;
