+++ src/compositions/pages/Home/BuildBetter/styled.js
@@ -1,6 +1,7 @@
 import styled from 'styled-components';
+import Container from '../../../../components/Container';
 import { highlight, text } from '../../../../styles/colours';
-import { gridWidth } from '../../../../styles/spacing';
+import { spacing, gridWidth } from '../../../../styles/spacing';
 
 export const Title = styled.h2`
   color: ${text};
@@ -11,6 +12,11 @@ export const Title = styled.h2`
     content: '.';
     color: ${highlight};
   }
+`;
+
+export const Section = Container.extend`
+  margin-bottom: ${spacing(10)};
+  margin-top: ${spacing(10)};
 `;
 
 export const Body = styled.p`
